+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The official release will have a defined and more stable API. If you are already
 * Improvements:
  * Display status in output at the end of a build step (#4)
  * Display total build time in history (#16)
+ * Indicate lost connection to LambdaCD in UI
 * API Changes:
  * removed deprecated argument lists for `lambdacd.internal.execution/execute-steps`